Add basic main.go to get things going
I plan on using GoKit, but I just have
some of those basics stubbed in at the
moment. At this point, I know my server
is working.

I spent a bit of time looking into setting
up a true secure connection locally,
but I ran into issues with not being
able to bind to the :443 or :80 ports
with ListenAndServeTLS. From what I could
find online, it looks like Linux disallows
binding to these ports with sudo or some
other capabilities. It looked like
a bit of a hassle to get this working
locally, so I decided to skip this for
now and get to the fun stuff of setting
up the web-scraper.