Add crypto-jobs-list endpoint w/ filtering
I moved the scraping to its own package
so I can easily import it in a cmd/
later on. The scraping will be handled
by some other automated process. The API
requests will pull data from the BoltDB which
I will populate with each scrape.

For this reason I also moved the filtering
to the http handler, but this is just temporary.
This filter query will be moved to BoltDB
next; loading the entire dataset into memory
to filter isn't a good idea.